pyproject.toml
[tool.poetry]
name = "sound-converter"
version = "0.1.0"
description = ""
authors = ["Z1pP <maxm_99@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14" 
pywifi = "^1.1.12"
comtypes = "^1.4.8"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
mypy = "^1.14.1"
types-requests = "^2.32.0.20241016"
types-beautifulsoup4 = "^4.12.0.20241020"
aiohttp = "^3.11.11"
aiofiles = "^24.1.0"
pyinstaller = "^6.11.1"
pyqt6 = "^6.8.0"
qasync = "^0.27.1"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "PyQt6.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aiohttp.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aiofiles.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "bs4.*"
ignore_missing_imports = true
